backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11
psycopg[binary]>=3.1
gunicorn>=21.2
uvicorn>=0.30
PyJWT>=2.9,<3.0
